refactor: 수익률 출력 로직을 BigDecimal 사용으로 수정
- double 타입을 사용하는 printEarningsRate 메서드를 BigDecimal 사용으로 변경함.
- 새로운 formatEarningsRate 메서드를 추가하여 수익률을 백분율로 포맷팅.
- DecimalFormat 대신 BigDecimal의 divide 메서드를 사용하여 수익률을 정확하게 계산하고 포맷팅함.

+++ src/main/java/lotto/controller/LottoController.java
@@ -106,7 +106,7 @@ public class LottoController {
         long totalPrizeMoney = prizeCalculator.getTotalPrizeMoney();
         BigDecimal earningsRate = LottoEarningCalculator.calculateEarningsRate(totalPrizeMoney, numberOfTickets, lottoPrice);
 
-        lottoOutput.printEarningsRate(earningsRate.doubleValue());
+        lottoOutput.printEarningsRate(earningsRate);
     }
 
     private List<String> prepareStatisticsLines(Map<Integer, Integer> prizeCountMap) {
